Add category update on products

+++ routes/categoryRouter.js
@@ -1,5 +1,6 @@
 const categoryRouter = require("express").Router();
 const Category = require("../models/Category");
+const Product = require("../models/Product");
 const verifyAdminToken = require("./verifyAdminToken");
 const verifyToken = require("./verifyToken");
 
@@ -40,13 +41,19 @@ categoryRouter.get("/title/:title", verifyToken, async (req, res) => {
 
 //
 // Update a category by id using the save mode to work
-// asynchronous and with validation
+// asynchronous and with validation.
+// Check if the title is different. If yes, the category title
+// field on the products must also change.
 //
 categoryRouter.put("/:id", verifyAdminToken, async (req, res) => {
   const updateCategory = await Category.findById(req.params.id);
   if (!updateCategory) {
     return res.status(400).send("Error getting category");
   }
+
+  // Save the actual category title in another variable
+  // to use in products after the validation check.
+  const checkCategoryTitle = updateCategory.title;
 
   updateCategory.title = req.body.title;
   updateCategory.description = req.body.description;
@@ -57,6 +64,22 @@ categoryRouter.put("/:id", verifyAdminToken, async (req, res) => {
   let error = updateCategory.validateSync();
   if (error) {
     return res.status(400).send(error);
+  }
+
+  // Check if there is an update on title.
+  // If there is a change, we must also update
+  // the category field on products.
+  if (checkCategoryTitle !== req.body.title) {
+    const updateProducts = await Product.updateMany(
+      { category: checkCategoryTitle },
+      { $set: { category: req.body.title } }
+    );
+
+    if (!updateProducts) {
+      return res
+        .status(400)
+        .send("Error updating category entries in products");
+    }
   }
 
   await updateCategory.save();
@@ -91,14 +114,32 @@ categoryRouter.post("/add_category", verifyAdminToken, async (req, res) => {
 });
 
 //
-// Delete a category by id
+// Delete a category by id.
+// First check if a product uses this category.
+// If yes, deleting the category is invalid.
+// The user must first delete the products
+// connected to this category.
 //
 categoryRouter.delete("/:id", verifyAdminToken, async (req, res) => {
-  const deleteCategory = await Category.deleteOne({ _id: req.params.id });
-  if (!deleteCategory) {
-    return res.status(400).send("Error deleting category");
+  // Get the category data for the delete request
+  const getCategory = await Category.findById(req.params.id);
+  if (!getCategory) {
+    return res.status(400).send("Error getting category");
   }
-  res.status(200).send("Deleting category was successful");
+
+  // Check if the category to delete has connected products
+  const getProduct = await Product.findOne({ category: getCategory.title });
+  // If not, we delete the category
+  if (!getProduct) {
+    const deleteCategory = await Category.deleteOne({ _id: req.params.id });
+    if (!deleteCategory) {
+      return res.status(400).send("Error deleting category");
+    }
+    res.status(200).send("Deleting category was successful");
+  } else {
+    // The category has connected products and we cannot delete it
+    return res.status(400).send("The category is not empty");
+  }
 });
 
 module.exports = categoryRouter;
